Added 1/2

+++ backend/database/migrations/2026_04_23_999998_rename_stock_to_product_stocks_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
             $table->foreignId('branch_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity')->default(0);
+            $table->decimal('quantity', 10, 2)->default(0);
             $table->integer('minimum_stock')->default(0);
             $table->timestamps();
